+++ EEprom.h
@@ -8,12 +8,12 @@ extern byte serialDebug;
 class EEprom
 {
 public:
-	EEprom(uint8_t addr);
-	bool write(uint16_t address, byte data);
-	byte read(uint16_t address);
+	EEprom(int addr);
+	bool write(int address, byte data);
+	byte read(int address);
 	void eraseAll();
 	void readAll();
-	uint16_t _Addr;
+	int _Addr;
 private:
 	
 };
